% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa.R
\name{apa.loading.table}
\alias{apa.loading.table}
\title{Generate table for the item loadings}
\usage{
apa.loading.table(x, table.number = NA, model = "cfa")
}
\arguments{
\item{x}{an object from the result of lavaan::cfa function.}

\item{table.number}{a number presenting the order of the table.}

\item{model}{a string which means a kind of analysis,such as "cfa","ESEM" or "ESEM within cfa"}
}
\value{
a table_all object
}
\description{
Generate table for the item loadings
}
