% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa.R
\name{apa.par.table}
\alias{apa.par.table}
\title{Generate table for parameter estimates}
\usage{
apa.par.table(
  data,
  par,
  table.number = NA,
  model = "ESEM within cfa",
  digits = NA
)
}
\arguments{
\item{data}{an object from the result of lavaan::cfa function.}

\item{par}{a list of parameters to have in the table.}

\item{table.number}{a number presenting the order of the table.}

\item{model}{a string which means a kind of analysis,such as "cfa" or "ESEM within cfa"}

\item{digits}{a number to keep the decimal number.}
}
\value{
a table_all object.
}
\description{
This function can generate a table list for the parameter estimates from 'cfa' function,
which can be used to make an APA table in following.
}
\examples{
\dontrun{
data2<-read.csv(file="D:/Users/squarrel/rdata/shuju.csv",header=true,sep=",")
model1<-'f1=~amo1+amo2+amo3+amo4'
fit<-lavaan::cfa(model1,data2)
apa.par.table(fit,c("cfi","tli"),table.number=1,model="cfa",digits=2)
}
}
