% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mori.R
\name{ci.col}
\alias{ci.col}
\title{A tool to make a confidence interval column}
\usage{
ci.col(x)
}
\arguments{
\item{x}{an object from the result of get_load function.}
}
\value{
a data.frame
}
\description{
A tool to make a confidence interval column
}
