% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tab.R
\name{create_tab}
\alias{create_tab}
\title{Create a table in APA style}
\usage{
create_tab(
  x,
  latentname = NULL,
  itemname = NULL,
  title = NULL,
  digits = 3,
  model = NULL,
  type = "flextable"
)
}
\arguments{
\item{x}{an object from the result made by apa.loading.table function.}

\item{latentname}{a list of the latent variable names.}

\item{itemname}{a list of the item names made by itemname function.}

\item{title}{set a title for your table.}

\item{digits}{a number to keep the decimal number when choosing the "flextable" type.}

\item{model}{a string which means a kind of analysis,such as "cfa","esem"or "esem within cfa".}

\item{type}{type of the table format, choices are "flextable" and "gt".}
}
\value{
an HTML table.
}
\description{
This is a wrapper around "flextable" and "gt" package to create a table.
"Digits" in this function is only used when selecting a "flextable" type.
}
\examples{
\dontrun{
la<-c("f1","f2")
item<-list(f1=c("amo1","amo2","amo3"),f2=c("ex1","ex2","ex3"))
data2<-read.csv(file="D:/Users/squarrel/rdata/shuju.csv",header=true,sep=",")
model1<-'f1=~amo1+amo2+amo3+amo4'
fit<-lavaan::cfa(model1,data2)
mytable<-apa.loading.table(fit,1)
create_tab(mytable,la,item,title="ESEM loading",model="esem",type="flextable")
}
}
