% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mori.R
\name{esem_load}
\alias{esem_load}
\title{Get the normal ESEM loading output generated by esemComp::esem_efa function}
\usage{
esem_load(x)
}
\arguments{
\item{x}{a object from the result of esemComp::esem_efa function.}
}
\value{
a matrix
}
\description{
Although ESEM is fundamentally a confirmatory analysis, but we can also do this
in an exploratory way.The result of esem_efa function may not give the significant
level of item loadings.
}
