% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esemchart-package.R
\docType{package}
\name{esemchart-package}
\alias{esemchart}
\alias{esemchart-package}
\title{esemchart: Generate Tables for ICM-CFA or ESEM-within-CFA model}
\description{
ESEM incorporates more flexible EFA models into estimation by allowing items to cross-load on non-target factors.By choosing target rotation method,ESEM can estimate in a confirmatory way.Thus,ESEM will get less biased factor correlations.This package based on the work of Mateus et al--'esemComp'package. It can make ESEM in EFA way and also make a complexable ESEM within CFA model.In this package, we will make a tool to corporate with 'esemComp'package to generate tables for ESEM.You can just run your model into 'esemComp'package,then put the result into this package to generate tables for parameters or loadings.
}
\author{
\strong{Maintainer}: Shuyu Li \email{lishuyuna@126.com}

Authors:
\itemize{
  \item He Xiao
  \item Yingying Luo
  \item Jiahua Lei
}

}
\keyword{internal}
