% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mori.R
\name{get_load}
\alias{get_load}
\title{Get standardized loadings}
\usage{
get_load(x)
}
\arguments{
\item{x}{an object from lavaan::cfa function.}
}
\value{
a data.frame
}
\description{
Get standardized loadings
}
