% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mori.R
\name{itemname}
\alias{itemname}
\title{Get the item names which load onto their priori latent factors depending on the
input column numbers in the raw data}
\usage{
itemname(list, data)
}
\arguments{
\item{list}{a list of column numbers indicating the indexes of the items related.}

\item{data}{the data in which the variable names match the item numbers.}
}
\value{
a list object.
}
\description{
Get the item names which load onto their priori latent factors depending on the
input column numbers in the raw data
}
\examples{
\dontrun{
data2<-read.csv(file="D:/Users/squarrel/rdata/shuju.csv",header=true,sep=",")
itemname(list(f1=c(1:4),f2=c(5:8)),data2)
}
}
