% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mori.R
\name{main.load}
\alias{main.load}
\title{Get main loadings of the items when estimating an ESEM within cfa model}
\usage{
main.load(latentname, itemname, fit)
}
\arguments{
\item{latentname}{a list of latent variable names you would like to choose.}

\item{itemname}{a list of item name in raw data you would like to use.}

\item{fit}{an object from the result of lavaan::cfa function.}
}
\value{
a main_load object
}
\description{
Get main loadings of the items when estimating an ESEM within cfa model
}
\examples{
\dontrun{
list<-list(c("amo1","amo2","amo3","amo4"),c("ex1","ex2","ex3","ex4"))
data2<-read.csv(file="D:/Users/squarrel/rdata/shuju.csv",header=true,sep=",")
model1<-'efa("efa")*f1+
         efa("efa")*f2=~amo1+amo2+amo3+amo4'
fit<-lavaan::cfa(model1,data2)
main.load(c("f1","f2"),list,fit)
}
}
